Adding logout
Auth seems to break when restarting the server.

+++ crates/core/migrations/20240713224437_users.up.sql
@@ -1,6 +1,6 @@
 -- Add up migration script here
 Create Table If Not Exists users (
-  id BigInt Primary Key Generated Always As Identity,
+  id Integer Primary Key Generated Always As Identity,
   uuid Uuid Not Null Unique,
   username Text Not Null Unique,
   password Text Not Null
